% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{annotate}
\alias{annotate}
\title{Perform annotation (combined call to search_db and ranking)}
\usage{
annotate(pks, db, tol = 5, m = "pos", rand_af = F, db_ref = db)
}
\arguments{
\item{pks}{peak matrix of the MSI experiment. Usually obtained from rMSIproc}

\item{db}{database to search}

\item{tol}{tolerance in ppm. Default is 5ppm.}

\item{m}{ionization mode ("pos" or "neg"). Default is positive ion mode ("pos").}

\item{rand_af}{boolean indicating whether to use random adducts and fragments (for target-decoy validation method 1). During normal operation it should be set to False. Default False.}

\item{db_ref}{reference db (for target-decoy validation method 2). Default is db (No alternative reference db)}
}
\value{
a data.frame of class "rMSILipidAnnot". With the results of annotation.
}
\description{
Perform annotation (combined call to search_db and ranking)
}
\examples{
Missing examples

}
