% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{loadData}
\alias{loadData}
\title{Load fragmentation and adduct data}
\usage{
loadData()
}
\value{
A list containing lipis, lipiddict, adducts, fragmentation pos, fragmentation neg, and losses
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
loadData()

}
