% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ranking}
\alias{ranking}
\title{Add extra ranking scores}
\usage{
ranking(res, pks, m)
}
\arguments{
\item{res}{data.frame of class "rMSILipidAnnot". Usually obtained from rMSIfragment::search_db().}

\item{pks}{peak matrix of the MSI experiment. Usually obtained from rMSIproc}

\item{m}{ionization mode ("pos" or "neg"). Default is positive ion mode ("pos").}
}
\value{
a data.frame of class "rMSILipidAnnot". With updated ranking scores.
}
\description{
Add extra ranking scores
}
\examples{
Missing examples

}
