% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{search_db}
\alias{search_db}
\title{Search adducts and fragments in database}
\usage{
search_db(mz, db, tol = 5, m = "pos", rand_af = F, db_ref = db)
}
\arguments{
\item{mz}{list of masses in the dataset}

\item{db}{database to search}

\item{tol}{tolerance in ppm. Default is 5ppm.}

\item{m}{ionization mode ("pos" or "neg"). Default is positive ion mode ("pos").}

\item{rand_af}{boolean indicating whether to use random adducts and fragments (for target-decoy validation method 1). During normal operation it should be set to False. Default False.}

\item{db_ref}{reference db (for target-decoy validation method 2). Default is db (No alternative reference db)}
}
\value{
a data.frame of class "rMSILipidAnnot". Containing information about all hits. Including theoretical mz, molecular formula, compound name, lipid family and the ranking scores.
}
\description{
Search adducts and fragments in database
}
\examples{
Missing examples

}
